+++ test/input_output/AbstractAlgebra.jl
@@ -6,7 +6,7 @@ using AbstractAlgebra, Test, Primes, Groebner
     @test Groebner.groebner([R(0), R(0)]) == [R(0)]
     @test Groebner.groebner([R(0), R(3), R(0)]) == [R(1)]
 
-    for ground in [GF(2^31 - 1), GF(2^62 + 135), QQ]
+    for ground in [GF(Int64(2)^31 - 1), GF(2^62 + 135), QQ]
         for gb_ord in [Groebner.Lex(), Groebner.DegLex(), Groebner.DegRevLex()]
             R, x = polynomial_ring(ground, "x")
             @test Groebner.groebner([x^2 - 4, x + 2], ordering=gb_ord) == [x + 2]
@@ -32,7 +32,7 @@ end
     aa_orderings_to_test = [:lex, :degrevlex, :deglex]
     aa_grounds_to_test = [
         AbstractAlgebra.GF(2^62 + 135),
-        AbstractAlgebra.GF(2^31 - 1),
+        AbstractAlgebra.GF(Int64(2)^31 - 1),
         AbstractAlgebra.GF(17),
         AbstractAlgebra.QQ
     ]
